requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
psycopg2==2.9.7
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
nltk==3.8.1
beautifulsoup4==4.12.2
selenium==4.11.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
transformers==4.30.2
torch==2.0.1
spacy==3.6.0
